envs/recall_env/server/requirements.txt
openenv-core[core]>=0.2.2
sentence-transformers
numpy
pyyaml
fastapi>=0.115.0
uvicorn>=0.24.0
pydantic
